GB-50 feat: 통계 및 크롤링 준비
- 페어링과 맥주에 대한 AOP 로직 준비하고 있습니다.
- 맥주 위시리스트 기능 구현 및 서비스로직 추상화 완료했습니다.
- 로그인 정보가 필요한 곳에 검증 로직을 추가했습니다.
- 맥주 카테고리와 유저 다대다 연결 완료했습니다.

+++ be/src/main/java/be/domain/beer/service/BeerServiceImpl.java
@@ -1,5 +1,6 @@
 package be.domain.beer.service;
 
+import static be.domain.user.entity.enums.Role.*;
 import static be.global.config.CacheConstant.*;
 
 import java.util.ArrayList;
@@ -24,6 +25,9 @@ import be.domain.beer.repository.MonthlyBeerRepository;
 import be.domain.beercategory.entity.BeerCategory;
 import be.domain.beercategory.service.BeerCategoryService;
 import be.domain.beertag.entity.BeerTag;
+import be.domain.user.entity.User;
+import be.domain.user.entity.enums.Role;
+import be.domain.user.service.UserService;
 import be.global.exception.BusinessLogicException;
 import be.global.exception.ExceptionCode;
 import lombok.RequiredArgsConstructor;
@@ -32,6 +36,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class BeerServiceImpl implements BeerService {
 
+	private final UserService userService;
 	private final BeerRepository beerRepository;
 	private final BeerQueryRepository beerQueryRepository;
 	private final MonthlyBeerRepository monthlyBeerRepository;
@@ -44,6 +49,12 @@ public class BeerServiceImpl implements BeerService {
 	@Transactional
 	public Beer createBeer(Beer beer) {
 
+		// User loginUser = userService.getLoginUser();
+		//
+		// if (!loginUser.getRoles().contains(ROLE_ADMIN)) {
+		// 	throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
+		// }
+
 		Beer savedBeer = Beer.builder().build();
 
 		savedBeer.create(beer);
@@ -55,6 +66,12 @@ public class BeerServiceImpl implements BeerService {
 	@Override
 	@Transactional
 	public Beer updateBeer(Beer beer, Long beerId) {
+
+		User loginUser = userService.getLoginUser();
+
+		if (!loginUser.getRoles().contains(ROLE_ADMIN)) {
+			throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
+		}
 
 		Beer findBeer = findVerifiedBeer(beerId);
 
@@ -71,7 +88,13 @@ public class BeerServiceImpl implements BeerService {
 	@Transactional
 	public void deleteBeer(Long beerId) {
 
-		beerRepository.deleteById(beerId);
+		User loginUser = userService.getLoginUser();
+
+		if (!loginUser.getRoles().contains(ROLE_ADMIN)) {
+			throw new BusinessLogicException(ExceptionCode.FORBIDDEN);
+		} else {
+			beerRepository.deleteById(beerId);
+		}
 	}
 
 	//    public Beer isWishListedBeer(Beer beer, User user){
@@ -107,9 +130,10 @@ public class BeerServiceImpl implements BeerService {
 	@Override
 	@Transactional
 	public Beer getBeer(Long beerId) {
+
 		Beer findBeer = findVerifiedBeer(beerId);
-		findBeer.getBeerDetailsStatistics().addDailyViewCount();
-		return beerRepository.save(findBeer);
+
+		return findBeer;
 	}
 
 	@Override
